+++ src/dis.c
@@ -57,7 +57,7 @@ static inline int print_single_ins(FklVM *vm,
     case FKL_OP_BVEC:
     case FKL_OP_BOX:
     case FKL_OP_HASH:
-        CB_FMT("::%s", fklGetSubOpcodeName(op, ins_arg.ix));
+        CB_FMT("::%s", fklGetSubOpcodeName(op, (int8_t)ins_arg.ix));
         break;
     case FKL_OP_LOAD_LIB:
         CB_FMT("%" PRIu64, ins_arg.ux);
